vibe: アニメーション設定を設定ダイアログからパレットポップオーバーに移動
